Expanded expression handling, and more
Added or consolodated support for most assignment operators,
logical operators and comma expressions.  Also replaced assign_stmt
with expr_stmt in the process.

See the README for more details of changes.  Also fixed a few
mistakes in the README from before (ie some expressions weren't
recursive ... it was late and I was mixing BNF and EBNF in my head)