Added parens for 'if'and 'while' and updated tests
I purposely left errors in a couple but haven't updaded the .expected
files to the error output.  All the others are correct.  I'll keep these
tests because there's no reason not to but I'll either have to make a
lot of changes or write a bunch of new tests because they only test the
simple language of the original project.